requirements.txt
asgiref==3.2.7
Babel==2.8.0
boto3==1.14.12
botocore==1.17.12
cycler==0.10.0
Django==3.0.7
django-filter==2.3.0
django-phonenumber-field==4.0.0
django-storages==1.9.1
docutils==0.15.2
gunicorn==20.0.4
jmespath==0.10.0
kiwisolver==1.2.0
matplotlib==3.2.1
numpy==1.18.2
opencv-python==4.2.0.34
pandas==1.0.3
phonenumbers==8.12.4
Pillow==7.1.2
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
s3transfer==0.3.3
scipy==1.4.1
seaborn==0.10.0
six==1.14.0
sqlparse==0.3.1
termcolor==1.1.0
urllib3==1.25.9
whitenoise==5.1.0
